[CALCITE-6123] DruidAdapterIT#testInterleaveBetweenAggregateAndGroupOrderByOnMetrics fails
Running the following queries in the order specified below leads to
wrong results for Q2.

* Q1
```sql
select "store_state", "brand_name", "A"
from (
  select "store_state", "brand_name", sum("store_sales") + sum("store_cost") as A
  from "foodmart"
  group by "store_state", "brand_name")
order by "brand_name", "store_state"
limit 5
```

* Q2
```sql
select "store_state", "brand_name", "A"
from (
  select "store_state", "brand_name", sum("store_sales") - sum("store_cost") as A
  from "foodmart"
  group by "store_state", "brand_name")
order by "A"
limit 5
```
The queries resemble a lot and the problem appears only when they are
executed in a given order which means that probably there is a bug
in the Druid implementation of the cache.

Disable the query cache on historical nodes to avoid the wrong results
issue till the problem is identified and fixed.